[FEAT] Rebase 좋아요수, 좋아요갯수 연동

+++ Android/app/src/main/java/com/ssafy/kkaddak/domain/usecase/market/GetAllNftsUseCase.kt
@@ -12,7 +12,7 @@ import javax.inject.Singleton
 class GetAllNftsUseCase @Inject constructor(
     private val marketRepository: MarketRepository
 ){
-    suspend operator fun invoke(lastId: Int, limit: Int, onlySelling: Boolean): Resource<List<NftItem>> =
+    suspend operator fun invoke(lastId: Long, limit: Long, onlySelling: Boolean): Resource<List<NftItem>> =
         withContext(Dispatchers.IO) {
             marketRepository.getAllNfts(lastId, limit, onlySelling)
         }
